fix: full card cover uses correct theme
