+++ libs/journeys/ui/src/components/Card/FullCardCover/index.ts
@@ -1,0 +1,1 @@
+export { FullCardCover } from './FullCardCover'
